// src/app/users/utils/models/kin.ts
export interface Kin {
    _id: string;
    userId: string;
    email: string;
    phone: string;
    firstname: string;
    lastname: string;
    createdAt: string;
    dob: string;
    street: string;
    city: string;
    state: string;
    country: string;
    gender: string;
    image: string;
    relationship: string;
}